+++ platformio/commands/package.py
@@ -17,9 +17,13 @@ import tempfile
 from datetime import datetime
 
 import click
+from tabulate import tabulate
 
 from platformio import fs
+from platformio.clients.account import AccountClient
 from platformio.clients.registry import RegistryClient
+from platformio.exception import UserSideException
+from platformio.package.manifest.parser import ManifestParserFactory
 from platformio.package.meta import PackageSpec, PackageType
 from platformio.package.pack import PackagePacker
 from platformio.package.unpack import FileUnpacker, TARArchiver
@@ -79,26 +83,94 @@ def package_pack(package, output):
     default=True,
     help="Notify by email when package is processed",
 )
-def package_publish(package, owner, released_at, private, notify):
-    # publish .tar.gz instantly without repacking
-    if not os.path.isdir(package) and isinstance(
+@click.option(
+    "--non-interactive",
+    is_flag=True,
+    help="Do not show interactive prompt",
+)
+def package_publish(  # pylint: disable=too-many-arguments, too-many-locals
+    package, owner, released_at, private, notify, non_interactive
+):
+    click.secho("Preparing a package...", fg="cyan")
+    owner = owner or AccountClient().get_logged_username()
+    do_not_pack = not os.path.isdir(package) and isinstance(
         FileUnpacker.new_archiver(package), TARArchiver
-    ):
-        response = RegistryClient().publish_package(
-            package, owner, released_at, private, notify
-        )
-        click.secho(response.get("message"), fg="green")
-        return
-
+    )
+    archive_path = None
     with tempfile.TemporaryDirectory() as tmp_dir:  # pylint: disable=no-member
-        with fs.cd(tmp_dir):
-            p = PackagePacker(package)
-            archive_path = p.pack()
-            response = RegistryClient().publish_package(
-                archive_path, owner, released_at, private, notify
+        # publish .tar.gz instantly without repacking
+        if do_not_pack:
+            archive_path = package
+        else:
+            with fs.cd(tmp_dir):
+                p = PackagePacker(package)
+                archive_path = p.pack()
+
+        type_ = PackageType.from_archive(archive_path)
+        manifest = ManifestParserFactory.new_from_archive(archive_path).as_dict()
+        name = manifest.get("name")
+        version = manifest.get("version")
+        data = [
+            ("Type:", type_),
+            ("Owner:", owner),
+            ("Name:", name),
+            ("Version:", version),
+        ]
+        click.echo(tabulate(data, tablefmt="plain"))
+
+        # look for duplicates
+        _check_duplicates(owner, type_, name, version)
+
+        if not non_interactive:
+            click.confirm(
+                "Are you sure you want to publish the %s %s to the registry?\n"
+                % (
+                    type_,
+                    click.style(
+                        "%s/%s@%s" % (owner, name, version),
+                        fg="cyan",
+                    ),
+                ),
+                abort=True,
             )
+
+        response = RegistryClient().publish_package(
+            owner, type_, archive_path, released_at, private, notify
+        )
+        if not do_not_pack:
             os.remove(archive_path)
-            click.secho(response.get("message"), fg="green")
+        click.secho(response.get("message"), fg="green")
+
+
+def _check_duplicates(owner, type, name, version):  # pylint: disable=redefined-builtin
+    items = (
+        RegistryClient()
+        .list_packages(filters=dict(types=[type], names=[name]))
+        .get("items")
+    )
+    if not items:
+        return True
+    # duplicated version by owner
+    if any(
+        item["owner"]["username"] == owner and item["version"]["name"] == version
+        for item in items
+    ):
+        raise UserSideException(
+            "The package `%s/%s@%s` is already published in the registry"
+            % (owner, name, version)
+        )
+    other_owners = [
+        item["owner"]["username"]
+        for item in items
+        if item["owner"]["username"] != owner
+    ]
+    if other_owners:
+        click.secho(
+            "\nWarning! A package with the name `%s` is already published by the next "
+            "owners: `%s`\n" % (name, ", ".join(other_owners)),
+            fg="yellow",
+        )
+    return True
 
 
 @cli.command("unpublish", short_help="Remove a pushed package from the registry")
@@ -119,9 +191,9 @@ def package_publish(package, owner, released_at, private, notify):
 def package_unpublish(package, type, undo):  # pylint: disable=redefined-builtin
     spec = PackageSpec(package)
     response = RegistryClient().unpublish_package(
+        owner=spec.owner or AccountClient().get_logged_username(),
         type=type,
         name=spec.name,
-        owner=spec.owner,
         version=str(spec.requirements),
         undo=undo,
     )